+++ contracts/mocks/EtherReceiverMock.sol
@@ -3,9 +3,9 @@
 pragma solidity ^0.8.0;
 
 contract EtherReceiverMock {
-    uint private constant _ACCEPTETHER_FALSE = 1;
-    uint private constant _ACCEPTETHER_TRUE = 2;
-    uint private _acceptEther = _ACCEPTETHER_FALSE;
+    uint256 private constant _ACCEPTETHER_FALSE = 1;
+    uint256 private constant _ACCEPTETHER_TRUE = 2;
+    uint256 private _acceptEther = _ACCEPTETHER_FALSE;
 
     function setAcceptEther(bool acceptEther) public {
         _acceptEther = acceptEther ? _ACCEPTETHER_TRUE : _ACCEPTETHER_FALSE;
